[WIP] split poc into multiple modules.
