+++ lib/checkr_api_error.py
@@ -1,0 +1,4 @@
+
+
+class CheckerAPIError(RuntimeError):
+    pass
