+++ lib/checkr_api_failed_creation.py
@@ -1,0 +1,6 @@
+from lib.checkr_api_error import CheckerAPIError
+
+
+class CheckerAPIFailedCreation(CheckerAPIError):
+    pass
+
